Now using short instead of int for a lot fo stuff, may change back in the future depending on required functionality

+++ Connect Four/Interfaces/IPlayer.cs
@@ -1,8 +1,9 @@
 ﻿using Connect_Four.Classes;
+using Connect_Four.Classes.GameBoard;
 
 namespace Connect_Four.Interfaces;
 
 public interface IPlayer
 {
-    int MakeMove(Board board);
+    short MakeMove(Board board);
 }
